Do not always close tab when leaving view 'TestDetails' via navigation
When leaving the view 'TestDetails', the associated tab is closed. This
behavior is unfortunately bad as an intentionally opened tab is closed
after starting a recording without finishing it, for example.

The behavior of closing the tab is only wanted for a record replay,
where the tab is opened automatically. The

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>