Properly identify UI5 control for selected DOM element without JQuery
The function '_getControlFromDOM' now properly identifies the UI5
control from a given DOM element – without using JQuery. For this, the
DOM tree is traverse up until an element with either of the attributes
'data-sap-ui' or 'data-sap-ui-related' is found, from which the control
can be identified.

The functionality is adapted from the code in the UI5 JQuery function in
'sap/ui/dom/jquery/control-dbg.js.

Documentation for this function and related parts of the code is added
or extended.

Signed-off-by: Claus Hunsen <claus.hunsen@msg.group>